Add files via upload
In this Application I use Visual studio to run code and Python Version is 3.11.10
Python provides us with many libraries to create GUI(Graphical User Interface), and Tkinter is one of them. Creating GUI with Tkinter is very easy and it is the most commonly used library for GUI development. In this article, we will create a Quiz application using Tkinter. A Quiz application has a set of questions and their answers and it checks for the correctness of answers that are submitted by users.
Points to remember while creating a basic GUI window.
1.Create Tkinter variables to get answers submitted by a user and to show the score last.
2.Create a Label to Give Heading to our GUI.
3.Create a button to start the quiz. When the user will click this button start_quiz function will be called.
4.Define a frame, this frame will be used to show questions and take user responses.
5.Define a button to show the next question but don’t pack it as we don’t want to see it until the user starts the quiz.
6.import required modules.
from tkinter import *
7.Define the start_quiz function. This function is used to remove the start quiz button and render the next question button. It also calls the next_question function to render the question.
8.Define the check_ans function. This function is used to check the answer that is submitted by the user and update the score.